@@ -27,7 +27,7 @@ class LoginPageState extends State<LoginPage> {
 
   Future<bool> apiLogin() async {
     if (_email == null || _password == null) {
-      if (context.mounted) {
+      if (mounted) {
         await MyAlertDialog.showErrorAlertDialog(
           key: const Key('login-alertdialog_emptyfields'),
           context: context,
@@ -64,7 +64,7 @@ class LoginPageState extends State<LoginPage> {
         return true;
       } else {
         if (jsonData.containsKey('message')) {
-          if (context.mounted) {
+          if (mounted) {
             printServerResponse(context, response, 'apiLogin',
                 message: jsonData['message']);
             return false;
@@ -72,7 +72,7 @@ class LoginPageState extends State<LoginPage> {
         }
       }
     } catch (err, stacktrace) {
-      if (context.mounted) {
+      if (mounted) {
         printCatchError(context, err, stacktrace,
             message: AppLocalizations.of(context)!.connectionRefused);
         setState(() {
@@ -80,6 +80,7 @@ class LoginPageState extends State<LoginPage> {
         });
         return false;
       }
+      return false;
     }
     return false;
   }
@@ -138,7 +139,10 @@ class LoginPageState extends State<LoginPage> {
         printCatchError(context, err, stacktrace,
             message:
                 AppLocalizations.of(context)!.errorOccurredDuringPasswordReset);
+
+        return;
       }
+      return;
     }
   }
 
