@@ -1,18 +1,18 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
+import 'package:http/http.dart' as http;
 import 'package:provider/provider.dart';
+import 'package:risu/components/alert_dialog.dart';
 import 'package:risu/components/appbar.dart';
 import 'package:risu/components/burger_drawer.dart';
+import 'package:risu/components/loader.dart';
 import 'package:risu/components/outlined_button.dart';
+import 'package:risu/globals.dart';
 import 'package:risu/pages/article/article_list_data.dart';
 import 'package:risu/pages/home/home_page.dart';
 import 'package:risu/pages/rent/confirm/confirm_rent_page.dart';
-import 'package:risu/utils/providers/theme.dart';
-import 'package:risu/components/alert_dialog.dart';
-import 'package:risu/components/loader.dart';
-import 'package:risu/globals.dart';
-import 'package:http/http.dart' as http;
 import 'package:risu/utils/errors.dart';
+import 'package:risu/utils/providers/theme.dart';
 
 class ConfirmRentState extends State<ConfirmRentPage> {
   late int hours;
@@ -36,7 +36,7 @@ class ConfirmRentState extends State<ConfirmRentPage> {
         _loaderManager.setIsLoading(false);
       });
       if (response.statusCode == 201) {
-        if (context.mounted) {
+        if (mounted) {
           await MyAlertDialog.showInfoAlertDialog(
             context: context,
             title: AppLocalizations.of(context)!.invoiceSent,
@@ -44,14 +44,14 @@ class ConfirmRentState extends State<ConfirmRentPage> {
           );
         }
       } else {
-        if (context.mounted) {
+        if (mounted) {
           printServerResponse(context, response, 'sendInvoice',
               message: AppLocalizations.of(context)!
                   .errorOccurredDuringSendingInvoice);
         }
       }
     } catch (err, stacktrace) {
-      if (context.mounted) {
+      if (mounted) {
         setState(() {
           _loaderManager.setIsLoading(false);
         });
